% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_case_when.R
\docType{methods}
\name{add_case_when.DBIConnection}
\alias{add_case_when.DBIConnection}
\alias{show,CustomTranslation-method}
\title{Add a custom SQL translation to a DBI connection}
\usage{
\method{add_case_when}{DBIConnection}(con, ...)

\S4method{show}{CustomTranslation}(object)
}
\arguments{
\item{con}{A DBIConnection object.}

\item{...}{Not used.}

\item{object}{A customised connection.}
}
\value{
A new DBIConnection object.
}
\description{
Add a custom SQL translation to a DBI connection
}
