% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customised_translation.R
\name{add_custom_translation.DBIConnection}
\alias{add_custom_translation.DBIConnection}
\title{Add a custom SQL translation to a DBI connection}
\usage{
\method{add_custom_translation}{DBIConnection}(object, ...)
}
\arguments{
\item{object}{A DBIConnection object.}

\item{...}{Not used.}
}
\value{
A new DBIConnection object.
}
\description{
Add a custom SQL translation to a DBI connection
}
