% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customised_translation.R
\name{add_custom_translation}
\alias{add_custom_translation}
\title{Add a customised translation to an object}
\usage{
add_custom_translation(object, ...)
}
\arguments{
\item{object}{An object.}

\item{...}{Other arguments.}
}
\description{
Add a customised translation to an object
}
