% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sql_translate.R
\name{add_sql_translate}
\alias{add_sql_translate}
\title{Register a case_when function in a SQL translator}
\usage{
add_sql_translate(cw, con = NULL)
}
\arguments{
\item{cw}{A case_when function.}

\item{con}{A connection.}
}
\description{
Register a case_when function in a SQL translator
}
