% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sql_translate.R
\name{add_sql_translate2.case_when}
\alias{add_sql_translate2.case_when}
\title{Register a case_when function}
\usage{
\method{add_sql_translate2}{case_when}(fn, con = NULL, ...)
}
\description{
Register a case_when function
}
