% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sql_translate.R
\name{add_sql_translate2}
\alias{add_sql_translate2}
\title{Register a function in a SQL translator}
\usage{
add_sql_translate2(fn, con = NULL, ...)
}
\arguments{
\item{fn}{A function to register.}

\item{con}{A connection.}

\item{...}{Other arguments passed on to method.}
}
\description{
Register a function in a SQL translator
}
