% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_variant.R
\name{add_sql_variant}
\alias{add_sql_variant}
\title{Register a case_when function in a SQL translator}
\usage{
add_sql_variant(cw, con = NULL)
}
\arguments{
\item{cw}{A case_when function.}

\item{con}{A connection.}
}
\description{
Register a case_when function in a SQL translator
}
