% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casewhen-package.R
\docType{package}
\name{casewhen-package}
\alias{casewhen}
\alias{casewhen-package}
\title{casewhen: Reusable "case_when" Functions}
\description{
With 'casewhen' package, you can create 'dplyr::case_when()' functions
that might be use many times.
}
\author{
\strong{Maintainer}: Romain Lesur \email{romain.lesur@gmail.com}

}
\keyword{internal}
