% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_case_when.R
\name{create_case_when}
\alias{create_case_when}
\alias{formulas.case_when}
\alias{variable.names.case_when}
\alias{print.case_when}
\title{A case_when factory}
\usage{
create_case_when(..., vars = "x")

\method{formulas}{case_when}(x, ...)

\method{variable.names}{case_when}(object, ...)

\method{print}{case_when}(x, ...)
}
\arguments{
\item{...}{For \code{formulas}, \code{variable.names} and \code{print}: not used.}

\item{vars}{A character vector that determined the names of the arguments
of the returned function.}

\item{x, object}{A \code{case_when} function, created by \code{create_case_when}.}
}
\value{
A function, usable in place of \code{\link[dplyr:case_when]{dplyr::case_when()}}.
}
\description{
\code{create_case_when} allows to create reusable \code{\link[dplyr:case_when]{dplyr::case_when()}} functions.
It returns a function that can be used in place of
\code{\link[dplyr:case_when]{dplyr::case_when()}}. The arguments of the returned function are
determined by the \code{vars} argument of \code{create_case_when()}.
}
\details{
The returned function is of class \code{case_when}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{formulas}: Get the formulas of a \code{case_when} function.

\item \code{variable.names}: Get the variable names of a \code{case_when} function.

\item \code{print}: Print informations about a \code{case_when} function.
}}

\examples{
x <- 1:50
y <- 51:100

cw_fb <- create_case_when(
  number \%\% 35 == 0 ~ "fizz buzz",
  number \%\% 5 == 0 ~ "fizz",
  number \%\% 7 == 0 ~ "buzz",
  TRUE ~ as.character(number),
  vars = "number"
)

cw_fb(number = x)
cw_fb(number = y)

# Formulas and variable names can be extracted
patterns <- formulas(cw_fb)
var_name <- variable.names(cw_fb)

# Dots support splicing
create_case_when(!!! patterns, vars = var_name)
}
