% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_case_when.R
\name{formulas}
\alias{formulas}
\title{Get formulas}
\usage{
formulas(x, ...)
}
\arguments{
\item{x}{An object used to select a method.}

\item{...}{Other arguments passed on to methods.}
}
\value{
A list of \code{formula} objects.
}
\description{
\code{formulas} retrieves formulas from an object that depends on.
}
\keyword{internal}
