% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_case_when.R
\docType{methods}
\name{show,CustomTranslation-method}
\alias{show,CustomTranslation-method}
\title{Show method for a CustomTranslation object}
\usage{
\S4method{show}{CustomTranslation}(object)
}
\arguments{
\item{object}{A customised connection.}
}
\description{
Show method for a CustomTranslation object
}
\keyword{internal}
