% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_case_when.R
\name{add_case_when.DBIConnection}
\alias{add_case_when.DBIConnection}
\title{Add a custom SQL translation to a DBI connection}
\usage{
\method{add_case_when}{DBIConnection}(con, ...)
}
\arguments{
\item{con}{A DBIConnection object.}

\item{...}{Not used.}
}
\value{
A new DBIConnection object.
}
\description{
Add a custom SQL translation to a DBI connection
}
