% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_case_when.R
\name{add_case_when}
\alias{add_case_when}
\title{Add a customised translation to a connection}
\usage{
add_case_when(con, ...)
}
\arguments{
\item{con}{A database connection.}

\item{...}{Other arguments.}
}
\description{
Add a customised translation to a connection
}
