% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_case_when.R
\name{add_case_when}
\alias{add_case_when}
\title{Add a custom SQL translation to a DBI connection}
\usage{
add_case_when(con, ...)
}
\arguments{
\item{con}{A DBIConnection object.}

\item{...}{Not used.}
}
\value{
A new DBIConnection object.
}
\description{
Add a custom SQL translation to a DBI connection
}
