% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sql_translate.R
\name{add_sql_translate}
\alias{add_sql_translate}
\alias{add_sql_translate.case_when}
\title{Register a function in a SQL translator}
\usage{
add_sql_translate(fn, con = NULL, ...)

\method{add_sql_translate}{case_when}(fn, con = NULL, ...)
}
\arguments{
\item{fn}{A function to register.}

\item{con}{A database connection.}

\item{...}{Other arguments passed on to method.}
}
\description{
Register a function in a SQL translator
}
\section{Methods (by class)}{
\itemize{
\item \code{case_when}: Register a case_when function in a SQL
translate environment
}}

