% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casewhen-package.R
\docType{package}
\name{casewhen-package}
\alias{casewhen}
\alias{casewhen-package}
\title{casewhen: Create Reusable 'dplyr' 'case_when()' Functions}
\description{
Helpers to avoid repetition of similar 'case_when()' functions when
performing data wrangling with the 'dplyr' package. It also works with the
'dbplyr' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RLesur/casewhen#readme}
  \item Report bugs at \url{https://github.com/RLesur/casewhen/issues}
}

}
\author{
\strong{Maintainer}: Romain Lesur \email{romain.lesur@gmail.com}

}
\keyword{internal}
