% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_case_when.R
\name{create_case_when}
\alias{create_case_when}
\title{A \code{case_when} Factory}
\usage{
create_case_when(..., vars = "x")
}
\arguments{
\item{...}{Arguments passed on to \code{dplyr::case_when}
\describe{
  \item{}{}
}}

\item{vars}{A character vector of generic variable names.}
}
\value{
A function, usable in place of \code{dplyr::case_when}.
}
\description{
This function allows to create reusable \code{dplyr::case_when()} functions.
}
\examples{
library(dplyr)

people <- tribble(
  ~name, ~sex, ~seek,
  "Mary", "F", "M",
  "Henry", "M", "F"
)

cw_sex <- create_case_when(x == "F" ~ "Woman",
                           x == "M" ~ "Man",
                           TRUE ~ as.character(x),
                           vars = c("x"))

people \%>\%
  mutate(sex_label = cw_sex(sex), seek_label = cw_sex(seek))
}
