% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_case_when.R
\name{create_sql_case_when}
\alias{create_sql_case_when}
\title{Create a SQL translation of a case_when function}
\usage{
create_sql_case_when(..., vars = "x", con = NULL)
}
\arguments{
\item{...}{A sequence of thwo sided formulas. The left hand side (LHS)
determines which values match this case.The right hand side (RHS)
provides the replacement value.

The LHS must evaluate to a logical vector. Each logical vector can
either have length 1 or a common length. All RHSs must evaluate to
the same type of vector.

These dots are evaluated with \href{dplyr::tidy-dots}{explicit splicing}.}

\item{vars}{A character vector.}

\item{con}{A database connection.}
}
\description{
Create a SQL translation of a case_when function
}
\keyword{internal}
